+++ example.php
@@ -1,7 +1,7 @@
 <?php
 error_reporting(E_ALL);
 ini_set('display_errors', 'On');
-require 'bnc.php';
+require 'Sbnc.php';
 $sbnc = new Sbnc\Sbnc();
 $sbnc->check();
 ?>
